Refactor tests to make it runnable within Google.
