+++ tests/common.go
@@ -38,8 +38,6 @@ import (
 	"log"
 	"net/http"
 	"os"
-	"os/exec"
-	"path/filepath"
 	"testing"
 	"time"
 
@@ -50,8 +48,6 @@ import (
 
 	"golang.org/x/crypto/ssh"
 	"golang.org/x/net/context"
-	"golang.org/x/oauth2"
-	"golang.org/x/oauth2/google"
 	compute "google.golang.org/api/compute/v1"
 )
 
@@ -72,10 +68,7 @@ var (
 	token          = flag.String("token", "", "When set, the proxy uses this Bearer token for authorization.")
 )
 
-const (
-	defaultOS       = "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20160329"
-	buildShLocation = "cmd/cloud_sql_proxy/build.sh"
-)
+const defaultOS = "https://www.googleapis.com/compute/v1/projects/debian-cloud/global/images/debian-8-jessie-v20160329"
 
 type logger interface {
 	Log(args ...interface{})
@@ -157,31 +150,6 @@ type process struct {
 // what's wrong.
 func (p *process) Close() error {
 	return p.sess.Close()
-}
-
-func compileProxy() (string, error) {
-	// Find the 'build.sh' script by looking for it in cwd, cwd/.., and cwd/../..
-	var buildSh string
-
-	var parentPath []string
-	for parents := 0; parents < 2; parents++ {
-		cur := filepath.Join(append(parentPath, buildShLocation)...)
-		if _, err := os.Stat(cur); err == nil {
-			buildSh = cur
-			break
-		}
-		parentPath = append(parentPath, "..")
-	}
-	if buildSh == "" {
-		return "", fmt.Errorf("couldn't find %q; please cd into the local repository", buildShLocation)
-	}
-
-	cmd := exec.Command(buildSh)
-	if out, err := cmd.CombinedOutput(); err != nil {
-		return "", fmt.Errorf("error during build.sh execution: %v;\n%s", err, out)
-	}
-
-	return "cloud_sql_proxy", nil
 }
 
 // startProxy executes the cloud_sql_proxy via ssh. The returned ReadCloser
@@ -408,24 +376,6 @@ func sshKey() (pubKey string, auth ssh.AuthMethod, err error) {
 		return "", nil, err
 	}
 	return string(ssh.MarshalAuthorizedKey(pub)), ssh.PublicKeys(signer), nil
-}
-
-func clientFromCredentials(ctx context.Context) (*http.Client, error) {
-	if f := *credentialFile; f != "" {
-		all, err := ioutil.ReadFile(f)
-		if err != nil {
-			return nil, fmt.Errorf("invalid json file %q: %v", f, err)
-		}
-		cfg, err := google.JWTConfigFromJSON(all, proxy.SQLScope)
-		if err != nil {
-			return nil, fmt.Errorf("invalid json file %q: %v", f, err)
-		}
-		return cfg.Client(ctx), nil
-	} else if tok := *token; tok != "" {
-		src := oauth2.StaticTokenSource(&oauth2.Token{AccessToken: tok})
-		return oauth2.NewClient(ctx, src), nil
-	}
-	return google.DefaultClient(ctx, proxy.SQLScope)
 }
 
 func TestMain(m *testing.M) {
